Fix #155 SpellVisualKitModelAttach with AttachmentId -1
When a spell visual kit model attachment with an AttachmentId of -1 was selected it would throw a hard erorr in the UI and fail to load the attachment.

This fixes the error, but does not fix the root cause of the issue. The database schema currently has the wrong data type, it stores it as a uint rather than a int.

+++ SpellGUIV2/MainWindow.xaml.cs
@@ -3366,7 +3366,7 @@ namespace SpellEditor
             ToggleVisualAttachments(attachRecord != null);
             if (attachRecord != null)
             {
-                VisualAttachmentIdCombo.SelectedIndex = int.Parse(attachRecord["AttachmentId"].ToString());
+                VisualAttachmentIdCombo.SelectedIndex = (int)uint.Parse(attachRecord["AttachmentId"].ToString());
                 VisualAttachmentOffsetXTxt.Text = attachRecord["OffsetX"].ToString();
                 VisualAttachmentOffsetYTxt.Text = attachRecord["OffsetY"].ToString();
                 VisualAttachmentOffsetZTxt.Text = attachRecord["OffsetZ"].ToString();
